ci: fix build-all flag
